move offset token to rewrite package
